Finish the SilhouetteDemo
Now, only the silhouette/outline of the mesh is drawn as lines.

The original source http://prideout.net/blog/?p=54 had an issue with
working in homogeneous space when determining the visibility/frontfacing
of a face, which was wrong under perspective projection. We need to do
the perspective divide to correct it.