Add multisampled edge detection demo.
This demo uses EXT_framebuffer_multisample and EXT_framebuffer_blit to
render the normals to a multisampled renderbuffer and blit/resolve it to
a single-sampled FBO with an attached texture.
That texture is afterwards used by the edge detection filter.
So, the shader itself does not work on a multisampled texture, and
therefore this demo is also suited for drivers/cards that do not support
multisampled textures via ARB_texture_multisample.