Add transform feedback raytracing demo
This demo shows how the OpenGL Transform Feedback mechanism can be used
to construct the scene geometry information for the Compute Shader as
Shader Storage Buffer Object. The benefit of using transform feedback is
that the vertices can be transformed normally in a vertex shader as done
with rasterization - including any model transformations - and then the
view-space vertices can be used directly by the ray tracer.
Otherwise, the ray tracer would have needed to perform model
transformations itself.