Add small optimization to stackless kd tree traversal not in the paper
When using kd tree traversal (stackless or stackful) we can abort
traversal when we find the first triangle hit in a leaf node. This is
because we always traverse the kd tree along the ray direction,
front-to-back.
It's just that with our current kd-tree construction in Java we do not
split triangles that belong to multiple leaf nodes. So we need to check
whether the hit point of the triangle was actually within the bounds of
the leaf node.