Compute not the shortest distance to the black hole center
But use the vector coplanar with the plane defined by the black hole
directed towards the camera.
This is because the projected quad lies on that plane and if we want the
distortion computations to not be off when we are too close to the black
hole, we need to adjust to the fact that we are using that projected
quad.