Add note on how to obtain pdf(w) from pdf(a)
pdf(w) = pdf(a) * dist^2 / cos(w)

where pdf(a) is 1/A and w = angle between surface normal and sample dir.
Since with the disk, the distance between the origin and the sampled
disk point is elongated by 1/cos(w), the two necessary cos(w) factors in
(d*d) cancel out, leaving one cos(w) (as 'c') in the denominator.