Use 16-bit signed normalized integers for positions and 32-bit indices
This is the only 16-bit integer type supported for ray tracing geometry
and the 16-bit floating point conversion has some precision issues at
values larger than 256, even though float16 should support all integers
in the range 0..2048 exactly.
Since it is a normalized type (in the range -1..1), we have to scale the
positions according to the number of bits we want to use for the
positions.