Add VKUtil.pointersOfElements()
when wanting to use more than one
VkAccelerationStructureBuildRangeInfoKHR in
vkCmdBuildAccelerationStructuresKHR we need to create a PointerBuffer
holding the addresses of all of the
VkAccelerationStructureBuildRangeInfoKHR structs.
Even though these examples only use one such struct, it is still clearer
this way what is meant by that code and easier to adapt it to multiple
structs.

This will be in LWJGL 3.3.2.